Refactor test utilities and language server tests: improve structure, remove unused code, and enhance readability.
