+++ jac/jaclang/langserve/tests/server_test/test_lang_serve.py
@@ -1,163 +1,20 @@
-"""Test suite for Jac language server features with improved structure and maintainability."""
+"""Test suite for Jac language server features."""
 
-import os
-from typing import Optional
-from dataclasses import dataclass
 import pytest
 
 from lsprotocol.types import (
-    DidOpenTextDocumentParams,
-    TextDocumentItem,
-    DidSaveTextDocumentParams,
-    DidChangeTextDocumentParams,
     DocumentFormattingParams,
     TextEdit,
-    VersionedTextDocumentIdentifier,
     TextDocumentIdentifier,
 )
 from jaclang.langserve.tests.server_test.utils import (
-    create_temp_jac_file,
-    load_jac_template,
+    TestFile,
+    LanguageServerTestHelper,
     create_ls_with_workspace,
+    load_jac_template,
 )
 from jaclang.vendor.pygls.uris import from_fs_path
-from jaclang.langserve.engine import JacLangServer
-from jaclang.langserve.server import did_open, did_save, did_change, formatting
-
-
-@dataclass
-class TestFile:
-    """Encapsulates test file information and operations."""
-    
-    path: str
-    uri: str
-    code: str
-    version: int = 1
-    
-    @classmethod
-    def from_template(cls, template_name: str, content: str = "") -> "TestFile":
-        """Create a test file from a template."""
-        code = load_jac_template(cls._get_template_path(template_name), content)
-        temp_path = create_temp_jac_file(code)
-        return cls(
-            path=temp_path,
-            uri=from_fs_path(temp_path),
-            code=code,
-        )
-    
-    @staticmethod
-    def _get_template_path(file_name: str) -> str:
-        """Get absolute path to test template file."""
-        return os.path.abspath(
-            os.path.join(os.path.dirname(__file__), file_name)
-        )
-    
-    def cleanup(self):
-        """Remove temporary test file."""
-        if os.path.exists(self.path):
-            os.remove(self.path)
-    
-    def increment_version(self) -> int:
-        """Increment and return the version number."""
-        self.version += 1
-        return self.version
-
-
-class LanguageServerTestHelper:
-    """Helper class for language server testing operations."""
-    
-    def __init__(self, ls: JacLangServer, test_file: TestFile):
-        self.ls = ls
-        self.test_file = test_file
-    
-    async def open_document(self) -> None:
-        """Open a document in the language server."""
-        # content = code if code is not None else self.test_file.code
-
-        params = DidOpenTextDocumentParams(
-            text_document=TextDocumentItem(
-                uri=self.test_file.uri,
-                language_id="jac",
-                version=self.test_file.version,
-                text=self.test_file.code,
-            )
-        )
-        await did_open(self.ls, params)
-    
-    async def save_document(self, code: Optional[str] = None) -> None:
-        """Save a document in the language server."""
-        content = code if code is not None else self.test_file.code
-        version = self.test_file.increment_version()
-        
-        if code:
-            self._update_workspace(code, version)
-        
-        params = DidSaveTextDocumentParams(
-            text_document=TextDocumentItem(
-                uri=self.test_file.uri,
-                language_id="jac",
-                version=version,
-                text=content,
-            )
-        )
-        await did_save(self.ls, params)
-    
-    async def change_document(self, code: str) -> None:
-        """Change document content in the language server."""
-        version = self.test_file.increment_version()
-        self._update_workspace(code, version)
-        
-        params = DidChangeTextDocumentParams(
-            text_document=VersionedTextDocumentIdentifier(
-                uri=self.test_file.uri,
-                version=version
-            ),
-            content_changes=[{"text": code}],
-        )
-        await did_change(self.ls, params)
-    
-    def _update_workspace(self, code: str, version: int) -> None:
-        """Update workspace with new document content."""
-        self.ls.workspace.put_text_document(
-            TextDocumentItem(
-                uri=self.test_file.uri,
-                language_id="jac",
-                version=version,
-                text=code,
-            )
-        )
-    
-    def get_diagnostics(self) -> list:
-        """Get diagnostics for the current document."""
-        return self.ls.diagnostics.get(self.test_file.uri, [])
-    
-    def get_semantic_tokens(self):
-        """Get semantic tokens for the current document."""
-        return self.ls.get_semantic_tokens(self.test_file.uri)
-    
-    def assert_no_diagnostics(self) -> None:
-        """Assert that there are no diagnostics."""
-        diagnostics = self.get_diagnostics()
-        assert isinstance(diagnostics, list)
-        assert len(diagnostics) == 0, f"Expected no diagnostics, found {len(diagnostics)}"
-    
-    def assert_has_diagnostics(self, count: int = 1, message_contains: Optional[str] = None) -> None:
-        """Assert that diagnostics exist with optional message validation."""
-        diagnostics = self.get_diagnostics()
-        assert isinstance(diagnostics, list)
-        assert len(diagnostics) == count, f"Expected {count} diagnostic(s), found {len(diagnostics)}"
-        
-        if message_contains:
-            assert message_contains in diagnostics[0].message, \
-                f"Expected '{message_contains}' in diagnostic message"
-    
-    def assert_semantic_tokens_count(self, expected_count: int) -> None:
-        """Assert semantic tokens data has expected count."""
-        tokens = self.get_semantic_tokens()
-        assert hasattr(tokens, "data")
-        assert isinstance(tokens.data, list)
-        assert len(tokens.data) == expected_count, \
-            f"Expected {expected_count} tokens, found {len(tokens.data)}"
+from jaclang.langserve.server import formatting
 
 
 class TestLangServe:
@@ -187,7 +44,8 @@ class TestLangServe:
         """Test opening a Jac file with syntax error produces diagnostics."""
         test_file = TestFile.from_template(self.CIRCLE_TEMPLATE, "error")
         uri, ls = create_ls_with_workspace(test_file.path)
-        test_file.uri = uri
+        if uri:
+            test_file.uri = uri
         helper = LanguageServerTestHelper(ls, test_file)
         
         await helper.open_document()
@@ -214,11 +72,10 @@ class TestLangServe:
         
         # Introduce syntax error
         broken_code = load_jac_template(
-            TestFile._get_template_path(self.CIRCLE_TEMPLATE), 
+            test_file._get_template_path(self.CIRCLE_TEMPLATE), 
             "error"
         )
-        helper._update_workspace(broken_code, test_file.increment_version())
-        await helper.open_document()
+        await helper.change_document(broken_code)
         helper.assert_has_diagnostics(count=1)
         helper.assert_semantic_tokens_count(self.EXPECTED_CIRCLE_TOKEN_COUNT)
         
@@ -230,7 +87,8 @@ class TestLangServe:
         """Test saving a Jac file triggers appropriate diagnostics."""
         test_file = TestFile.from_template(self.CIRCLE_TEMPLATE)
         uri, ls = create_ls_with_workspace(test_file.path)
-        test_file.uri = uri
+        if uri:
+            test_file.uri = uri
         helper = LanguageServerTestHelper(ls, test_file)
         
         await helper.open_document()
@@ -239,7 +97,7 @@ class TestLangServe:
         
         # Save with syntax error
         broken_code = load_jac_template(
-            TestFile._get_template_path(self.CIRCLE_TEMPLATE),
+            test_file._get_template_path(self.CIRCLE_TEMPLATE),
             "error"
         )
         await helper.save_document(broken_code)
@@ -254,7 +112,8 @@ class TestLangServe:
         """Test changing a Jac file triggers diagnostics."""
         test_file = TestFile.from_template(self.CIRCLE_TEMPLATE)
         uri, ls = create_ls_with_workspace(test_file.path)
-        test_file.uri = uri
+        if uri:
+            test_file.uri = uri
         helper = LanguageServerTestHelper(ls, test_file)
         
         await helper.open_document()
@@ -276,9 +135,11 @@ class TestLangServe:
         test_file = TestFile.from_template(self.CIRCLE_TEMPLATE)
         uri, ls = create_ls_with_workspace(test_file.path)
         
+        from lsprotocol.types import FormattingOptions
+        
         params = DocumentFormattingParams(
-            text_document=TextDocumentIdentifier(uri=uri),
-            options={"tabSize": 4, "insertSpaces": True},
+            text_document=TextDocumentIdentifier(uri=uri or ""),
+            options=FormattingOptions(tab_size=4, insert_spaces=True),
         )
         edits = formatting(ls, params)
         
@@ -297,8 +158,11 @@ class TestLangServe:
         file2 = TestFile.from_template(self.GLOB_TEMPLATE, "error")
         
         uri1, ls = create_ls_with_workspace(file1.path)
-        file1.uri = uri1
-        file2.uri = from_fs_path(file2.path)
+        if uri1:
+            file1.uri = uri1
+        file2_uri = from_fs_path(file2.path)
+        if file2_uri:
+            file2.uri = file2_uri
         
         helper1 = LanguageServerTestHelper(ls, file1)
         helper2 = LanguageServerTestHelper(ls, file2)
@@ -317,7 +181,7 @@ class TestLangServe:
         
         # Change first file
         changed_code = load_jac_template(
-            TestFile._get_template_path(self.GLOB_TEMPLATE),
+            file1._get_template_path(self.GLOB_TEMPLATE),
             "glob x = 90;"
         )
         await helper1.change_document(changed_code)
